maitake::sync
Module mutex

An asynchronous mutual exclusion lock.

See the documentation on the Mutex type for details.

Structs

Lock

A future returned by the Mutex::lock method.

Mutex

An asynchronous mutual exclusion lock for protecting shared data.

MutexGuard

An RAII implementation of a “scoped lock” of a Mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.

OwnedMutexGuard

An RAII implementation of a “scoped lock” of a Mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.