Module maitake::sync::wait_map

A map of Wakers associated with keys, so that a task can be woken by key.

See the documentation for the WaitMap type for details.

Structs

EnqueueWait

A future that ensures a Wait has been added to a WaitMap.

Wait

Future returned from WaitMap::wait().

WaitMap

A map of Wakers associated with keys, allowing tasks to be woken by their key.

WaitOwned

Future returned from WaitMap::wait_owned().

Enums

WaitError

Errors returned by WaitMap::wait, indicating a failed wake.

WakeOutcome

The result of an attempted WaitMap::wake() operation.

Type Aliases

WaitResult

The result of a call to WaitMap::wait().