Module maitake::sync::semaphore

An asynchronous counting semaphore.

A semaphore limits the number of tasks which may execute concurrently. See the Semaphore type’s documentation for details.

Structs

Acquire

The future returned by the Semaphore::acquire method.

AcquireOwned

Future returned from Semaphore::acquire_owned().

OwnedPermit

An owned RAII guard representing one or more permits acquired from a Semaphore.

Permit

A RAII guard representing one or more permits acquired from a Semaphore.

Semaphore

An asynchronous counting semaphore.

Enums

TryAcquireError

Errors returned by Semaphore::try_acquire.